fix: 修复2022/10/24 单元测试出现异常
Description: 程序退出时停止“同步回收站与最近删除”线程

Log: 修复2022/10/24 单元测试出现异常
Bug: https://pms.uniontech.com/bug-view-166419.html

+++ tests/test_mainwindow.cpp
@@ -1561,7 +1561,9 @@ TEST(MainWindow, monitor)
         currentFile.copy(currentInfo.absoluteFilePath() + "." + currentInfo.suffix());
     }
 
-    QFile::remove(infosInPictures[0].absoluteFilePath());
+    if (infosInPictures.size() > 0) {
+        QFile::remove(infosInPictures[0].absoluteFilePath());
+    }
 
     QTest::qWait(5000);
